Specify node engine minimum version
Setting the minimum version of the node engine will make npm to warn or
complain, if node `engine-strict` config is set to true, whenever the
installation environment does not meet this requirement.

Background:

`css-loader` depends on `postcss` that uses native promises. Promises
where introduced in node version 0.12.0. So `css-loader`
*transitively inherits* this dependency.

Related to #144.