Manage loader resolution with Webpack. Fixes #286.
Instead of assuming that child-loaders are the same as the parent, use
Webpack’s build in dependency resolution. The benefits of this is that
it’s easier to treat the CSS dependency tree differently (i.e. extract
text for certain vendor libs). The tradeoff is that it gives people the
ability to shoot themselves in the foot if they misconfigure their
loaders. This is worthwhile IMHO as the error will fail fast.

It shouldn’t change behavior for existing users.