docs(readme): adding extract css chunks to readme
Adding it to the documentation as a alternative  extended version of mini css extract.
Extract css chunks offers hot module reloading while in development, allowing for good performance on the client and build.
As you guys will know, mini css is fast, this brings the benefits of it to development without losing the ability to hot reload.